+++ belle-sip/tester/belle_sip_register_tester.c
@@ -694,7 +694,7 @@ static belle_sip_request_t* send_message(belle_sip_request_t *initial_request, c
 }
 static void reuse_nonce(void) {
 	belle_sip_request_t *register_request;
-	int initial_auth_context_count=belle_sip_list_size(prov->auth_contexts);
+	size_t initial_auth_context_count=belle_sip_list_size(prov->auth_contexts);
 	register_request=register_user_at_domain(stack, prov, "tcp",1,"marie","sip.linphone.org",NULL);
 	if (register_request) {
 		belle_sip_header_authorization_t * h = NULL;
@@ -712,7 +712,7 @@ static void reuse_nonce(void) {
 		belle_sip_provider_add_sip_listener(prov,BELLE_SIP_LISTENER(listener));
 
 		/*currently only one nonce should have been used (the one for the REGISTER)*/
-		BC_ASSERT_EQUAL(belle_sip_list_size(prov->auth_contexts), initial_auth_context_count+1,int,"%d");
+		BC_ASSERT_EQUAL((unsigned int)belle_sip_list_size(prov->auth_contexts), (unsigned int)initial_auth_context_count+1,unsigned int,"%u");
 
 		/*this should reuse previous nonce*/
 		message_request=send_message(register_request, auth_domain);
